+++ resources/js/Components/Bet/EventCard.vue
