+++ resources/js/Pages/Football/Index.vue
@@ -4,7 +4,7 @@ import { Head } from '@inertiajs/vue3';
 import { useBets } from '@/scripts/stores/bets.js';
 import { storeToRefs } from 'pinia';
 import { onMounted, onUnmounted } from 'vue';
-import BetCard from '@/Components/Bet/BetCard.vue';
+import EventCard from '@/Components/Bet/EventCard.vue';
 
 const store = useBets();
 const { allBets } = storeToRefs(useBets());
@@ -37,7 +37,7 @@ onUnmounted(() => {
 
         <div class="py-12 mx-auto max-w-7xl sm:px-6 lg:px-8">
             <div class="grid sm:grid-cols-2 gap-8 p-4 md:w-full">
-                <BetCard
+                <EventCard
                     v-for="event in allBets"
                     :key="event.id"
                     icon="/images/nfl-logo.png"
